NoteControllerクラスをWorkflowManagerクラスに統合し、NoteControllerクラスを削除/initializeメソッドをWorkflowManagerクラスに統合/オプションによる条件分岐をWorkflowManagerクラスの各メソッドに持たせ、main.jsのワークフローを変更
